+++ Manager/MainWindow.xaml.cs
@@ -231,7 +231,10 @@ namespace Sheepy.Modnix.MainGUI {
          var p = new Paragraph();
          if ( ! App.Settings.MinifyLoaderPanel ) {
             p.Inlines.Add( new Bold( new Run( AppControl.LIVE_NAME ) ) );
-            p.Inlines.Add( $"\tVer {SharedGui.AppVer}{build}\r" );
+            p.Inlines.Add( $"\tVer {SharedGui.AppVer}{build}" );
+            if ( SharedGui.AppVer != null )
+               p.Inlines.Add( " w/ PPML 0.2" );
+            p.Inlines.Add( "\r" );
          }
          p.Inlines.Add( "Status: " );
          p.Inlines.Add( state );