+++ Manager/Resources/Changelog.md
@@ -3,6 +3,8 @@ Modnix Changelog
 # Version 3, in development
 
 * New: Show PPML version in app status.
+* New: Mod Actions - a new system that allows mods to declare Actions and/or to support them.
+* Fix: .NET fallback resolver now use correct path separator.  Should have only cosmetic difference.
 * Mod: Removed deprecated logger and config param from mod initialisers. Use API instead.
 
 # Version 2.4, 2020-05-10