+++ Manager/Resources/Changelog.md
@@ -4,11 +4,15 @@ Modnix Changelog
 
 * New: Show PPML version in app status.
 * New: Mod Actions - a new system that allows mods to declare Actions and/or to support them.
-* Fix: .NET fallback resolver now use correct path separator.  Should have only cosmetic difference.
 * Mod: Removed deprecated logger and config param from mod initialisers. Use API instead.
-* Mod: api_add now accepts parameter-less functions such as Func<string> or Action.
 * Log: Mod loading messages are now logged with the mod's logger.
 * Other: Mod phases other than SplashMod now depends on correct Harmony to function. SplashMods can still be loaded without Harmony.
+ 
+# Version 2.5, in development
+
+* Fix: .NET fallback resolver now use correct path separator.  Should have only cosmetic difference.
+* Mod: api_add now accepts parameter-less functions such as Func<string> or Action.
+* Mod: Avoids are now processed after Requires, in the resolve loop.
 
 # Version 2.4, 2020-05-10